Update to Inventory contract and tests
Added events to the contract and worked out the modifiers so they
behaver correctly. Set up tests with beforeEach and afterEach functions
to create a fresh contract and set up an event listener. The variable
log must be set to true to get event log output. Working on tests for
current functionality.